completed few func

+++ libft/ft_strncmp.c
@@ -14,14 +14,15 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int				i;
+	size_t				i;
 	unsigned char	*us1;
 	unsigned char	*us2;
 
 	i = 0;
 	us1 = (unsigned char *)s1;
 	us2 = (unsigned char *)s2;
-	while ((us1[i] || us2[i]) && (int)n > i)
+
+	while ((us1[i] || us2[i]) && n > i)
 	{
 		if (us1[i] != us2[i])
 			return (us1[i] - us2[i]);
